Separate installation and download directories
Some times (e.g. when running parallel gradle build - which uses separate JVMs)
it is conviniet to download elastic search distribution only one and then
on each JVM instantiate elasticsearch instance with different installation directory ports (possibly random).
This is was not possible if embedded-elastic uses the same base directory for downloads and installation as this
results in redownloading elastic search distribution on every JVM'm.

This commit adds possibility to set different download directory and set it outside of install directory.
Default is then installDir is the same as downloadDir