Install plugin with --batch option in Elasticsearch 5.x or later
Before Elasticsearch 6.3.0, if a plugin requires additional permissions,
and standard input is closed or there is no console, then the plugin is
installed without asking the user to acknowledge the additional
permissions. Elasticsearch 6.3.0 introduced a breaking change where it
waits for the user to acknowledge, unless the install command is run
with the --batch option. Elasticsearch 5.x introduced this option.